all minor bugs fixed completed the customer side app

+++ app/src/main/java/golhar/cocomo/zinger/model/ConfigurationModel.java
@@ -107,9 +107,9 @@ public class ConfigurationModel implements Parcelable {
             dest.writeInt(isDeliveryAvailable);
         }
         if (isOrderTaken == null) {
-            dest.writeByte((byte) 0);
+            dest.writeInt(0);
         } else {
-            dest.writeByte((byte) 1);
+            dest.writeInt(1);
             dest.writeInt(isOrderTaken);
         }
     }
